// src/apiV2/a7-service/model/projectPreviewDto.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */

export interface ProjectPreviewDto {
  /** ID проекта */
  id?: string;
  /** Название проекта */
  name?: string;
}