// src/apiV2/a7-service/model/userRolesItem.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */

export type UserRolesItem = 'admin' | 'owner' | 'agency' | 'cluster' | 'supervisor' | 'maker' | 'prompt' | 'remote';


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRolesItem = {
  admin: 'admin' as UserRolesItem,
  owner: 'owner' as UserRolesItem,
  agency: 'agency' as UserRolesItem,
  cluster: 'cluster' as UserRolesItem,
  supervisor: 'supervisor' as UserRolesItem,
  maker: 'maker' as UserRolesItem,
  prompt: 'prompt' as UserRolesItem,
  remote: 'remote' as UserRolesItem,
};